requirements.txt
annotated-types==0.7.0
anyio==4.8.0
boto3==1.36.1
botocore==1.36.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
fastapi==0.115.6
gunicorn==23.0.0
h11==0.14.0
idna==3.10
jmespath==1.0.1
mongoengine==0.29.1
packaging==24.2
pydantic==2.10.5
pydantic_core==2.27.2
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
s3transfer==0.11.1
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
urllib3 
uvicorn==0.34.0
passlib==1.7.4
bcrypt==3.2.0
itsdangerous
pyjwt
python-dotenv